perform_index_merge_for_test
